Extract validates_polymorph into module
