+++ lib/polymorpheus/interface.rb
@@ -17,24 +17,9 @@ module Polymorpheus
     end
 
     def self.included(base)
-      base.extend(ClassMethods)
       base.extend(BelongsToPolymorphic)
       base.extend(HasManyAsPolymorph)
+      base.extend(ValidatesPolymorph)
     end
-
-    module ClassMethods
-
-      def validates_polymorph(polymorphic_api)
-        validate Proc.new {
-          unless polymorpheus.active_association
-            association_names = polymorpheus.associations.map(&:name)
-            errors.add(:base, "You must specify exactly one of the following: "\
-                              "{#{association_names.join(', ')}}")
-          end
-        }
-      end
-
-    end
-
   end
 end
